Add mechanism to prefix all C exported symbols with internal_ prefix
This is done in provision for the integration of PROJ master with GDAL. GDAL is
a complex library that links with many libraries, a number of them linking with
PROJ. However in the continuous integration setups used by GDAL, it would be
impractical to recompile all those libraries that use the system libproj (4.X or 5.X)
that comes with the distribution. Linking GDAL directly against PROJ master
and this system libproj with potential different ABI is prone to clash/crash
(we experimented that painfully in GDAL with its internal libtiff 4.X copy whereas
systems shipped with libtiff 3.X)
Hence this solution to rename the symbols of PROJ master so that PROJ master
can be used by GDAL without conflicting with the system PROJ.
The renaming only happens if -DPROJ_RENAME_SYMBOLS is passed in CFLAGS and CXXFLAGS.

Note: we potentially should do the same for C++ symbols, but that is not needed
for now, and we would just set the NS_PROJ existing macro to change the C++ base
namespace which defaults to osgeo::proj currently.