Add proj_get_crs_info_list_from_database()
This method is intended to be used typically by GUI that lists all possible CRS.

What is does could be done by previously existing functions, but it is much faster.
It typically runs in less than 0.1s (hot run) versus ~0.5s with the method that
consists in enumerating all codes and instanciating a PJ object for each of them.