Reverse direction of deformation operations/transformations/deformation
Changed the direction of dt-calculation to follow the same convention as
helmert. Changed from dt = t_c - t_obs to dt = t_obs - t_c, which
effectively reverses the direction of the operation. Existing projstrings
using deformation can simply reverse the direction of the operation to
achieve the same results as before this commit.