Allow gie to use proj_create_crs_to_crs (#1314)
The commands crs_src and crs_dst are introduced to allow operation
initialization with proj_create_crs_to_crs. This has the benefit of
using proj.db to select the proper operation. Additionally axis order is
respected for the coordinate systems that are tested. This is beneficial
when running the GIGS test suite for instance.