ocea: fix behaviour when +alpha is specified
The one-point case was completely broken with lat_0 being ignored.
I've fixed that, and also modified the alpha orientation, so that the angle
corresponds to the one of the 2 point case when going from point 1 to point 2,
similarly to what omerc does.
This was found rather experimentaly with the added test cases that try to
find equivalence between 1-point and 2-point cases.

Fixes #1379
and adresses https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=13930