PROJ string ingestion: adds +towgs84= when expanding +init=epsg:XXXX (fixes #1597)
Currently, +init=epsg:XXXX expansion from proj.db didn't make any effort
to try to add a +towgs84 clause. Now this will do using CRS::createBoundCRSToWGS84IfPossible()
Note that however this does not perfectly emulated PROJ4/PROJ5 epsg file
behaviour, and will only add a +towgs84 when it is valid for the whole
area of use of the CRS, whereas PROJ4/PROJ5 would add the one that was
valid for the largest sub-area of use.
The additiong of a +towgs84 clause will be thus more likely for a
projected CRS valid for a country, rather than for a geographic CRS valid
for a continent or a larger area, and thus whose transformations to WGS84
might be multiple.
The consequence of this shows up in some GIGS test that must be adapted to
add an explicit +towgs84 for source/target geographic CRS with value equal
to the one of the source/target projected CRS.