Improve Visual Studio compatibility check in CMake.
Make the setting of major toolset version variables parallel that for
cross-compilation.  Add comments so it's easier for the next person
looking at this code to figure what's going on.  Note: no need to unset
variables at the end of the version check file (the file is read in a
nested scope).