doxygen build script: don't output to /tmp
The doxygen log-files was previously put in /tmp. This is usually fine
on UNIX-type systems but when building the docs on Windows with Cygwin/MSYS2
tools the script can fail since that path is guaranteed to exist. This
commit makes sure that the doxygen script outputs its files in a path
relative to the current working directory.