Speed-up inverse approximate tmerc for coordinates with same northing
This PR is on top of #2039 for convenience, but is conceputally independent
from it.

When gdalwarp is used to reproject an image to a tmerc projection, the inverse
tmerc projection is actually used, and called by batch of coordinates belonging
to the same line in the output image, that is lines of same northing. Looking
at formulas, a lot of intermediate computations, including the use of hyperbolic
functions, only depend on the northing value. So cache them to potentially speed
up following transformations.

I couldn't get a significant value for the extra cost of doing that caching in
situations where it is not needed ('random' points). It is likely below 1%
(which is largely compensated by the improvements of #2039)

By modifying my bench program, to simulate 10 times the processing of  a
2000x2000 image (so 2000 coordinates of same northing being transformed
consecutively), I get a 40% speed-up over the ones of PR #2039

Some timings with gdalwarp on a 4000x4000 Byte single-channel image in EPSG:32631:

$ time OSR_USE_APPROX_TMERC=YES gdalwarp test.tif out.tif -t_srs EPSG:32631 -et 0 -overwrite

With this PR :   3.15s (30% improvement over PR #2039, 41% improvement over master)
With PR #2039:   4.54s (15% improvement over master)
Before PR #2039: 5.39s