Replace carriage return in input line by null character.
As result of a modification in logging (adding a carriage return), many
changes had to made in different places of cct.cpp. However, one missed
the CR in input to cct.
As result of that, the output from cct showed an superfluous empty line
after each output line. Replacing the CR in "comment", the last part of
an input line, by the null character ('\0') solves this issue. (yes, an
empty comment still terminates w/ a CR in memory).

Modification in logging?
https://github.com/OSGeo/PROJ/commit/37da5e243191c04607597f6b8a77acfa017a5c99
( cct: revise end-of-line handling in logging, and always output debug
 ... )
commit 37da5e (20191205)

See also:
https://github.com/OSGeo/PROJ/issues/1677
( cct outputs excessive whitespace comments are included in input data
 #1677 )