More tests verifying the output of the cct utility
On request more tests have been added verifying the output of
the cct utility.

cct has an option (-d) which overrides the default value for
the number of decimals (precision) that are printed when cct
outputs the value of a coordinate.

By default the arguments to the printf format string will be:

 * radians, degrees: width 14, precision 10
 * meters: width 13, precision 4 [1]

[1] height will be shown with a width of 12.

Note: the 4th "coordinate", time, will always be shown with
a precision of 4 decimals (and width 12).

When overridden (using the -d option) all coordinates, but
not time, will be shown with the specified decimals.