Replace 'switch' with 'if' in dmstor_ctx().
In response to PR feedback from @rouault on pull #2791, there was an
unintentional fallthrough in the switch statement case for the degree
sign.

Replacing the 'switch' altogether removes the fallthrough risk, allows us
to check for both bytes in the degree sign in the same test, and avoids
the need to duplicate the code from the 'default' case in multiple
branches.