Replace figure on usage/ellipsoids document.
The author the original figure did not provide a source for the image.
The image also had "geoid" misspelled in its labels, and label placement
was somewhat confusing.

This new image was created from scratch to match the design intent of
the original image, but reuses none of its assets directly.  I include
the SVG source.  This image is licensed CC0.