Fix unreachable code in projections/isea.
The internal helper function isea_hex() had an early unconditional
return, which made most of the code in the function totally unreachable.

Based on the comments in the unreachable section, I speculate that the
original author meant for this early return to occur for aperture values
other than 3 or 4.

In passing, change the function's return type to void, since it was
always returning 1 anyway and the return was not used (per #947), and
clean up some rogue whitespace.

Fixes #423.