Adapt testing to recursive pkg-config requirements
When pkg-config is called with --static, it will recursively collect
all Libs and Libs.private which are discovered via Requires and
Requires.private. So these libs must be available for linking even
when acutally only lib proj uses static linkage.
Libs and pkg-config must be provided for the target triplet.