Refactor real horner code
As discussed in https://github.com/OSGeo/PROJ/pull/3133 the existing
functions are now a bit complicated and can be refactor to simpler ones.

Checks for invalid direction are now unnecessary, as the forward and
reverse paths are only accessible through fwd4d and inv4d function
callbacks in the projection object.

Checks for a nullptr HORNER transformation object are too unnecessary as
enough checks are done in the projection object construction.